requirements.txt
python-telegram-bot==20.3
python-dotenv
python-binance
nest_asyncio